refactor(servlet): move authentication related servlet to the corresponding package
